backend/requirements.txt
fastapi
uvicorn
requests
pandas
openpyxl
faiss-cpu
sentence-transformers
numpy
scikit-learn
pymongo
python-jose[cryptography]
passlib[bcrypt]
beautifulsoup4
resend
python-dotenv
pdfplumber
python-multipart
